requirements.txt
openai>=1.0.0
schedule>=1.2.0
pandas>=2.0.0
google-cloud-bigquery>=3.0.0
pyyaml>=6.0.0
requests>=2.31.0
twilio>=9.0.0
